modify role_serviceImpl on feat create new role, SecurityProperties, …

+++ src/main/java/com/example/demo/db/reposiroty/BrandRepository.java
@@ -7,5 +7,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
  * @author Sattya
  * create at 1/28/2024 5:29 PM
  */
-public interface BrandRepository extends JpaRepository<BrandEntity,Integer> {
+public interface BrandRepository extends JpaRepository<BrandEntity,Long> {
 }
