+++ src/main/java/com/example/demo/db/reposiroty/CategoryRepository.java
@@ -7,5 +7,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
  * @author Sattya
  * create at 1/27/2024 3:50 PM
  */
-public interface CategoryRepository extends JpaRepository<CategoryEntity,Integer> {
+public interface CategoryRepository extends JpaRepository<CategoryEntity,Long> {
 }
